refactor(frontend): extract duplicated helpers into shared modules
Several small utilities and one large chunk of analytics chart logic
were copy-pasted across multiple pages, drifting slightly with each
copy. Extract them into shared/lib and shared/ui/analytics modules
so there's one implementation to maintain.
