+++ app/frontend/shared/lib/formatUsage.ts
@@ -1,0 +1,12 @@
+export function formatCostCents(cents: number): string {
+  const d = (cents ?? 0) / 100;
+  if (d >= 1000) return `$${(d / 1000).toFixed(1)}k`;
+  return `$${d.toFixed(2)}`;
+}
+
+export function formatTokens(n: number): string {
+  if (!n) return '0';
+  if (n >= 1_000_000) return `${(n / 1_000_000).toFixed(1)}M`;
+  if (n >= 1_000) return `${(n / 1_000).toFixed(1)}k`;
+  return String(n);
+}
